requirements.txt
Pillow
numpy
opencv-python
pytest
